Revert "Initial layout"
