--- client/public/docs/pexels-picjumbocom-196639.jpg:Zone.Identifier
@@ -1,4 +1,0 @@
-[ZoneTransfer]
-ZoneId=3
-ReferrerUrl=https://www.pexels.com/
-HostUrl=https://images.pexels.com/photos/196639/pexels-photo-196639.jpeg?cs=srgb&dl=pexels-picjumbocom-196639.jpg&fm=jpg
