--- client/src/components/Button.css
@@ -1,5 +1,0 @@
-
-.button {
-  background-color: tomato;
-  
-}
