--- client/src/components/Button.jsx
@@ -1,9 +1,0 @@
-import React, { className } from "react";
-
-export default function Button(props) {
-  return (
-    <button className="button">
-      {props.children}
-    </button>
-  );
-}
