--- client/src/components/ReceiptItem.css
@@ -1,7 +1,0 @@
-.receipt-item {
-  box-shadow: 0 3px 9px 0 grey;
-  max-width: 20%;
-  min-width: 20em;
-  padding: 1em;
-  margin: 1em;
-}
