--- client/src/components/ReceiptList.css
@@ -1,13 +1,0 @@
-.receipt-list {
-  box-shadow: 0 3px 9px 0 grey;
-  padding: 5em;
-  width: 80%;
-  align-items: center;
-  gap: 1em;
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  align-items: center;
-  flex-wrap: wrap;
-
-}
