--- client/src/components/View.jsx
@@ -1,9 +1,0 @@
-import ReceiptList from './ReceiptList';
-
-export default function View() {
-  return (
-    <div className='view'>
-      <ReceiptList />
-    </div>
-  )
-}
