--- client/src/components/Welcome.jsx
@@ -1,9 +1,0 @@
-import React, { className } from "react";
-
-export default function Welcome() {
-  return (
-    <div className="welcome">
-      <img src="https://github.com/mdzwink/my-purchases/blob/initial-layout/client/public/docs/pexels-picjumbocom-196639.jpg?raw=true" width={'100%'} />
-    </div>
-  );
-}
