+++ src/App.css
@@ -38,36 +38,24 @@
 }
 
 .welcome {
-  width: 100%;
-  height: 500px;
-  overflow: hidden;
-  margin: 0px;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
 }
 
 .navbar {
-  position: fixed;
-  width: 100%;
   background-color: white;
   box-shadow: 0px 1px 9px 0px grey;
   display: flex;
   flex-direction: row;
   justify-content: space-between;
   align-items: center;
-  list-style-type: none;
-  padding: 0 20px 0 20px;
-  margin: 0px;
 }
 
 .menu-items {
   display: flex;
   flex-direction: row;
-  padding-right: 40px;
+  margin-right: 1em;
   gap: 0.5em;
-}
-
-.view {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: 100%;
 }
