+++ src/App.js
@@ -1,14 +1,12 @@
 import './App.css';
 import Navbar from './components/Navbar';
 import Welcome from './components/Welcome';
-import View from './components/View';
 
 function App() {
   return (
     <div className="App">
       <Navbar />
       <Welcome />
-      <View />
     </div>
   );
 }
