+++ src/components/Navbar.jsx
@@ -1,0 +1,13 @@
+
+
+export default function Navbar() {
+  return (
+      <ul className="navbar">
+        <li><h2>Purchases Hub</h2></li>
+        <div className="menu-items">
+          <li><button>Add a Purchase</button></li>
+          <li><button>My Purchases</button></li>
+        </div>
+      </ul>
+  )
+}
