+++ src/components/View.jsx
@@ -1,0 +1,14 @@
+
+
+export default function View() {
+  return (
+    <>
+      <ul>
+        <li>Receipt 1</li>
+        <li>Receipt 2</li>
+        <li>Receipt 3</li>
+        <li>Receipt 4</li>
+      </ul>
+    </>
+  )
+}
