+++ src/components/Welcome.jsx
@@ -1,0 +1,9 @@
+
+
+export default function Welcome() {
+  return (
+    <div className="welcome">
+      <h1>Welcome!</h1>
+    </div>
+  );
+}
